test: Added CAM unittest
